pyproject.toml
################################################################################
# PROJECT METADATA
################################################################################

[project]
name = "QS-Trader"
version = "0.2.0-beta.5"
description = "Quantitative Trading Environment"
readme = "README.md"
requires-python = ">=3.13"
license = { text = "MIT" }
authors = [{ name = "QS-Trader", email = "quant@quantspace.io" }]
maintainers = [{ name = "QS-Trader", email = "quant@quantspace.io" }]
keywords = [
    "trading",
    "backtesting",
    "quantitative-finance",
    "algorithmic-trading",
]
classifiers = [
    "Development Status :: 3 - Alpha",
    "Intended Audience :: Financial and Insurance Industry",
    "Intended Audience :: Developers",
    "License :: OSI Approved :: MIT License",
    "Operating System :: OS Independent",
    "Programming Language :: Python :: 3",
    "Programming Language :: Python :: 3.13",
    "Topic :: Office/Business :: Financial :: Investment",
    "Topic :: Scientific/Engineering",
    "Typing :: Typed",
]

################################################################################
# DEPENDENCIES
################################################################################

dependencies = [
    # Data Processing
    "pandas>=2.3.2",
    "pyarrow>=21.0.0",
    # CLI & UI
    "click>=8.0.0",
    "rich>=13.0.0",
    # Configuration & Validation
    "pydantic>=2.11.9",
    "pyyaml>=6.0",
    "jsonschema>=4.25.1",
    # Logging
    "structlog>=24.4.0",
    # Visualization
    "plotly>=5.0.0",
    # Market data
    "yfinance>=0.2.66",
]

[project.scripts]
qs-trader = "qs_trader.cli:main"

[dependency-groups]
dev = [
    # Code Quality
    "isort>=6.0.1",
    "ruff>=0.13.0",
    "mypy>=1.14.0",
    "pre-commit>=4.3.0",
    # Testing
    "pytest>=8.4.2",
    "pytest-cov>=6.0.0",
    "pytest-xdist>=3.8.0",
    "pytest-watch>=4.2.0",
    # Documentation
    "mdformat>=0.7.21",
    "mdformat-gfm>=0.3.6",
    "mdformat-tables>=1.0.0",
    "mdformat-frontmatter>=2.0.8",
    # Type Stubs
    "pandas-stubs>=2.3.2.250926",
    "types-pyyaml>=6.0.12.20250915",
    "types-requests>=2.32.4.20250913",
    "types-jsonschema>=4.25.1.20251009",
    "plotly-stubs>=0.0.6",
]

################################################################################
# BUILD SYSTEM
################################################################################

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[tool.hatch.build.targets.wheel]
packages = ["src/qs_trader"]

[tool.hatch.build.targets.wheel.force-include]
"src/qs_trader/scaffold" = "qs_trader/scaffold"
"src/qs_trader/templates" = "qs_trader/templates"
"src/qs_trader/services/data" = "qs_trader/services/data"

[tool.uv]
package = true # Enable package installation for entry points


################################################################################
# CODE FORMATTING & LINTING
################################################################################

[tool.ruff]
line-length = 120
target-version = "py313"
src = ["src"]

[tool.ruff.lint]
ignore = [
    "F541",
] # Allow f-strings without placeholders (useful for consistency)

[tool.ruff.format]
quote-style = "double"
indent-style = "space"
skip-magic-trailing-comma = false
line-ending = "auto"

[tool.isort]
profile = "black"
multi_line_output = 3
line_length = 120
include_trailing_comma = true
force_grid_wrap = 0
use_parentheses = true
ensure_newline_before_comments = true
src_paths = ["src"]

[tool.mdformat]
wrap = "no"        # Do not wrap lines
end_of_line = "lf" # Use LF line endings (Unix style)

################################################################################
# TYPE CHECKING
################################################################################

[tool.mypy]
python_version = "3.13"
warn_return_any = true
warn_unused_configs = true
disallow_untyped_defs = false
check_untyped_defs = false
warn_redundant_casts = true
warn_unused_ignores = true
warn_no_return = true
warn_unreachable = true
strict_equality = true
show_error_codes = true

# Per-module options for third-party libraries without stubs
[[tool.mypy.overrides]]
module = "duckdb"
ignore_missing_imports = true

[[tool.mypy.overrides]]
module = "pyarrow.*"
ignore_missing_imports = true

################################################################################
# TESTING
################################################################################

[tool.pytest.ini_options]
testpaths = ["tests"]
python_files = ["test_*.py"]
python_classes = ["Test*"]
python_functions = ["test_*"]
addopts = [
    "-ra",              # Show extra test summary info for all except passed
    "--strict-markers", # Raise error on unknown markers
    "--strict-config",  # Raise error on unknown config options
    "--showlocals",     # Show local variables in tracebacks
]

################################################################################
# CODE COVERAGE
################################################################################

[tool.coverage.run]
source = ["src"] # Only measure coverage for src/ directory
omit = [
    "*/tests/*",         # Exclude all test files
    "*/__pycache__/*",   # Exclude cache
    "*/site-packages/*", # Exclude dependencies
]
branch = true # Measure branch coverage

[tool.coverage.report]
precision = 0 # No decimal places in percentages
show_missing = true # Show line numbers of missing coverage
skip_covered = false # Show all files, even 100% covered
exclude_lines = [
    "pragma: no cover",
    "def __repr__",
    "if TYPE_CHECKING:",
    "raise AssertionError",
    "raise NotImplementedError",
    "if __name__ == .__main__.:",
    "@abstractmethod",
    "@abc.abstractmethod",
]

[tool.coverage.html]
directory = "htmlcov"
